+++ core/src/test/java/bisq/core/dao/state/DaoStateSnapshotServiceTest.java
@@ -19,6 +19,7 @@ package bisq.core.dao.state;
 
 import bisq.core.dao.governance.period.CycleService;
 import bisq.core.dao.monitoring.DaoStateMonitoringService;
+import bisq.core.dao.state.storage.DaoStateStorageService;
 
 import org.junit.Before;
 import org.junit.Test;
